+++ foxes/models/turbine_types/__init__.py
@@ -1,4 +1,6 @@
 from .PCt_file import PCtFile
 from .PCt_from_two import PCtFromTwo
+from .CpCt_file import CpCtFile
+from .CpCt_from_two import CpCtFromTwo
 from .null_type import NullType
 from .wsrho2PCt_two_files import WsRho2PCtTwoFiles